Move Field to a new file, update references
